Added the Clegg Integrator module
The Clegg Integrator is similar to the Take Back Half algorithm that
gained so much popularity in the Nothing But Net game, but designed for
position control systems (as opposed to continuous-output systems in
Take Back Half's case). It will most likely be suitable for short
position based driving, but further testing will determine its best use
cases.